Add a route for post creation
