+++ frontend/src/components/layout/MainLayout.tsx
@@ -1,11 +1,8 @@
 import { Outlet } from 'react-router-dom';
 
-import CreatePost from '@/src/pages/CreatePost';
-
 const MainLayout = () => {
   return (
     <>
-      <CreatePost />
       <Outlet />
     </>
   );
